smoke test set up to run from jenkins
I removed the parallel testing so that my tests would run, and the login tests are now running in intellJ. I made a smoke test with just the login tests and then tried to run it from jenkins but it didnt work. I noticed an error message in the Test report XMLs configuration - is this the issue?